+++ src/main/java/net/minecraft/block/BlockPotato.java
@@ -19,16 +19,13 @@ public class BlockPotato extends BlockCrops
         return Items.potato;
     }
 
-    /**
-     * Spawns this Block's drops into the World as EntityItems.
-     */
     public void dropBlockAsItemWithChance(World worldIn, BlockPos pos, IBlockState state, float chance, int fortune)
     {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, fortune);
 
         if (!worldIn.isRemote)
         {
-            if (state.getValue(AGE) >= 7 && worldIn.rand.nextInt(50) == 0)
+            if (((Integer)state.getValue(AGE)).intValue() >= 7 && worldIn.rand.nextInt(50) == 0)
             {
                 spawnAsEntity(worldIn, pos, new ItemStack(Items.poisonous_potato));
             }
