+++ src/main/java/net/minecraft/client/renderer/WorldRenderer.java
@@ -33,8 +33,6 @@ public class WorldRenderer
     public int vertexCount;
     private VertexFormatElement vertexFormatElement;
     private int vertexFormatIndex;
-
-    /** None */
     private boolean noColor;
     public int drawMode;
     private double xOffset;
@@ -108,14 +106,14 @@ public class WorldRenderer
 
         for (int k = 0; k < ainteger.length; ++k)
         {
-            ainteger[k] = k;
+            ainteger[k] = Integer.valueOf(k);
         }
 
         Arrays.sort(ainteger, new Comparator<Integer>()
         {
             public int compare(Integer p_compare_1_, Integer p_compare_2_)
             {
-                return Floats.compare(afloat[p_compare_2_], afloat[p_compare_1_]);
+                return Floats.compare(afloat[p_compare_2_.intValue()], afloat[p_compare_1_.intValue()]);
             }
         });
         BitSet bitset = new BitSet();
@@ -124,7 +122,7 @@ public class WorldRenderer
 
         for (int l1 = 0; (l1 = bitset.nextClearBit(l1)) < ainteger.length; ++l1)
         {
-            int i1 = ainteger[l1];
+            int i1 = ainteger[l1].intValue();
 
             if (i1 != l1)
             {
@@ -133,7 +131,7 @@ public class WorldRenderer
                 this.rawIntBuffer.get(aint);
                 int j1 = i1;
 
-                for (int k1 = ainteger[i1]; j1 != l1; k1 = ainteger[k1])
+                for (int k1 = ainteger[i1].intValue(); j1 != l1; k1 = ainteger[k1].intValue())
                 {
                     this.rawIntBuffer.limit(k1 * l + l);
                     this.rawIntBuffer.position(k1 * l);
@@ -163,7 +161,7 @@ public class WorldRenderer
 
             for (int j2 = 0; j2 < ainteger.length; ++j2)
             {
-                int k2 = ainteger[j2];
+                int k2 = ainteger[j2].intValue();
                 atextureatlassprite[j2] = this.quadSprites[k2];
             }
 
@@ -422,9 +420,6 @@ public class WorldRenderer
         }
     }
 
-    /**
-     * Takes in the pass the call list is being requested for. Args: renderPass
-     */
     public int getColorIndex(int p_78909_1_)
     {
         return ((this.vertexCount - p_78909_1_) * this.vertexFormat.getNextOffset() + this.vertexFormat.getColorOffset()) / 4;
@@ -491,9 +486,6 @@ public class WorldRenderer
         }
     }
 
-    /**
-     * Disabels color processing.
-     */
     public void noColor()
     {
         this.noColor = true;
