+++ src/main/java/net/minecraft/item/ItemReed.java
@@ -19,15 +19,12 @@ public class ItemReed extends Item
         this.block = block;
     }
 
-    /**
-     * Called when a Block is right-clicked with this Item
-     */
     public boolean onItemUse(ItemStack stack, EntityPlayer playerIn, World worldIn, BlockPos pos, EnumFacing side, float hitX, float hitY, float hitZ)
     {
         IBlockState iblockstate = worldIn.getBlockState(pos);
         Block block = iblockstate.getBlock();
 
-        if (block == Blocks.snow_layer && iblockstate.getValue(BlockSnow.LAYERS) < 1)
+        if (block == Blocks.snow_layer && ((Integer)iblockstate.getValue(BlockSnow.LAYERS)).intValue() < 1)
         {
             side = EnumFacing.UP;
         }
