+++ src/main/java/net/minecraft/nbt/NBTTagFloat.java
@@ -7,7 +7,6 @@ import net.minecraft.util.MathHelper;
 
 public class NBTTagFloat extends NBTBase.NBTPrimitive
 {
-    /** The float value for the tag. */
     private float data;
 
     NBTTagFloat()
@@ -19,9 +18,6 @@ public class NBTTagFloat extends NBTBase.NBTPrimitive
         this.data = data;
     }
 
-    /**
-     * Write the actual data contents of the tag, implemented in NBT extension classes
-     */
     void write(DataOutput output) throws IOException
     {
         output.writeFloat(this.data);
@@ -33,12 +29,9 @@ public class NBTTagFloat extends NBTBase.NBTPrimitive
         this.data = input.readFloat();
     }
 
-    /**
-     * Gets the type byte for the tag.
-     */
     public byte getId()
     {
-        return 5;
+        return (byte)5;
     }
 
     public String toString()
@@ -46,9 +39,6 @@ public class NBTTagFloat extends NBTBase.NBTPrimitive
         return "" + this.data + "f";
     }
 
-    /**
-     * Creates a clone of the tag.
-     */
     public NBTBase copy()
     {
         return new NBTTagFloat(this.data);
